small rename (#214678)
